reorganised file structure, posts and users will now be populated with data when server runs
